maint: fix wrong PATH in testsuite
The switch to non-recursive make in v4.2.2-129-g3b29bec used a wrong
directory name in the PATH construction of automake's TESTS_ENVIRONMENT
variable ('/src/' instead of '/sed').
This had no effect as each test shell script appended the ./sed/ path
independently. However, switching to perl-based tests - this will be
noticed as the just-built sed binary will not be in the PATH.

* testsuite/local.mk: Fix $PATH variable in test environment.