tests: export CONFIG_HEADER to test scripts
Before this change, the testsuite/inplace-selinux test fails in some
cases when SELinux is available.  The reason it fails is because the
require_selinux_ function in init.cfg expects the CONFIG_HEADER
environment variable to be defined, but in fact it is not defined.
Even though we do invoke AM_CONFIG_HEADER in configure.ac, and even
though the CONFIG_HEADER variable does get set in the resulting
Makefile, nothing actually exports the CONFIG_HEADER variable, so the
test is doomed to fail whenever it is not skipped.  Fix this by
exporting the variable.

* testsuite/local.mk (TESTS_ENVIRONMENT): Add CONFIG_HEADER to the
list of variables to export to the environment of the test scripts.

Reported by Vineet Jain <vkjain@gmail.com> in
https://bugs.gnu.org/36150 .  The fix was suggested by Timothy Sample
<samplet@ngyro.com> in the same bug report.